Extend globbing (#1)
Activates extending globbing when processing the `.slugcleanup` file.

Our use case is due to Rail’s importmaps and Hotwire. The JavaScript controller folders need to persist in `app/javascript`z In out set-up the `app/javascript` folder is remove right before Heroku slug compilation finishes. Since these Hotwire controller folders are unique in that they are the only ones that need to stick around, an exclude glob is desired. This is not available in the standard globbing options, so an extension is added to allow the option.